requirements.txt
fastapi
uvicorn[standard]
pydantic
prisma
httpx
twilio
python-dotenv
